+++ gui/brassica-gui/ruleshighlighter.h
@@ -12,15 +12,18 @@ public:
 
     void setCategories(QStringList categories, bool forceUpdate = false);
     void setHighlights(QList<int> highlights);
+    void setErrors(QList<int> errors);
 
 protected:
     void highlightBlock(const QString &text) override;
 
 private:
     QList<int> highlights;
+    QList<int> errors;
 
     QTextCharFormat categoryFormat;
     QTextCharFormat highlightFormat;
+    QTextCharFormat errorFormat;
 
     // NB. these two shouldn't get out of sync!
     QStringList m_categories;